CDK-421: Refactor kite-tools, extract parent POM.
This splits kite-tools into a parent POM and a runtime POM. The runtime
project is named kite-tools so that artifact names do not change.

This also changes the packaging to use the maven-assembly-plugin rather
than the maven-shade-plugin. The assembly plugin doesn't requires a
whitelist of artifacts to include and should be more maintainable. The
shaded or assembled artifact name has not changed.

This removes the broken java command to run the CLI when Hadoop can't be
found.