CDK-421: Add kite-tools-cdh5 module.
This adds a kite-tools assembly for CDH5 packaged as a tarball and a zip
file. The assemblies contain a lib folder with all runtime and provided
dependencies for CDH5 (managed by the kite-tools-cdh5 pom) and a shell
script in bin called kite-dataset that looks for ../lib/* dependencies.
The bin script uses KITE_CLASSPATH for the lib folder, which can be
overridden. It also adds HADOOP_CONF_DIR and HIVE_CONF_DIR to the
classpath if those variables are set and are directories.